Pulls profile image from Facebook and applies it to a user's profile

+++ FriendNGo/app/src/main/java/com/friendngo/friendngo/friendngo/ActivityNotifications.java
@@ -114,7 +114,7 @@ public class ActivityNotifications extends AppCompatActivity {
                                 activity.getString("sender_home_city"),
                                 activity.getString("sender_home_nationality"),
                                 activity.getString("sender_points"),
-                                activity.getLong("id")));
+                                activity.getInt("id")));
                     } catch (JSONException e) {
                         Log.w("GET REQUEST FAIL1: ", e.getMessage().toString());
                     }
